# Faster rendering for renderableOrbitalKepler