--- a/modules/space/rendering/renderableorbitalkepler.cpp
+++ b/modules/space/rendering/renderableorbitalkepler.cpp
@@ -269,7 +269,7 @@ RenderableOrbitalKepler::Appearance::Appearance()
     renderingModes.addOptions({
         { RenderingModeTrail, "Trails" },
         { RenderingModePoint, "Points" },
-        { RenderingModePointTrail , "Points+Trails" }
+        { RenderingModePointTrail , "Points and Trails" }
     });
     addProperty(renderingModes);
 
@@ -331,10 +331,6 @@ RenderableOrbitalKepler::RenderableOrbitalKepler(const ghoul::Dictionary& dict)
     }
     addPropertySubOwner(_appearance);
 
-    _path = p.path.string();
-    _path.onChange([this]() { updateBuffers(); });
-    addProperty(_path);
-
     _format = codegen::map<kepler::Format>(p.format);
 
     _startRenderIdx = p.startRenderIdx.value_or(0);
@@ -362,6 +358,10 @@ RenderableOrbitalKepler::RenderableOrbitalKepler(const ghoul::Dictionary& dict)
     _contiguousMode = p.contiguousMode.value_or(false);
     _contiguousMode.onChange([this]() { _updateDataBuffersAtNextRender = true; });
     addProperty(_contiguousMode);
+
+    _path = p.path.string();
+    _path.onChange([this]() { updateBuffers(); });
+    addProperty(_path);
 }
 
 void RenderableOrbitalKepler::initializeGL() {
@@ -377,8 +377,7 @@ void RenderableOrbitalKepler::initializeGL() {
            return global::renderEngine->buildRenderProgram(
                "OrbitalKeplerTrails",
                absPath("${MODULE_SPACE}/shaders/debrisVizTrails_vs.glsl"),
-               absPath("${MODULE_SPACE}/shaders/debrisVizTrails_fs.glsl"),
-               absPath("${MODULE_SPACE}/shaders/debrisVizTrails_gs.glsl")
+               absPath("${MODULE_SPACE}/shaders/debrisVizTrails_fs.glsl")
            );
        }
    );
@@ -440,9 +439,6 @@ void RenderableOrbitalKepler::render(const RenderData& data, RendererTasks&) {
         return;
     }
 
-    GLint* _si = _startIndex.data();
-    GLint* _ss = _segmentSize.data();
-
     const int selection = _appearance.renderingModes;
     const bool renderPoints = (
         selection == RenderingModePoint ||
@@ -454,6 +450,8 @@ void RenderableOrbitalKepler::render(const RenderData& data, RendererTasks&) {
     );
 
     if (renderPoints) {
+        calculateSegmentsForPoints(data);
+
         _pointProgram->activate();
         _pointProgram->setUniform(
             _uniformPointCache.modelTransform,
@@ -506,9 +504,9 @@ void RenderableOrbitalKepler::render(const RenderData& data, RendererTasks&) {
         glBindVertexArray(_vertexArray);
         glMultiDrawArrays(
             GL_LINE_STRIP,
-            _si,
-            _ss,
-            static_cast<GLsizei>(_startIndex.size())
+            _startIndexPoints.data(),
+            _segmentSizePoints.data(),
+            static_cast<GLsizei>(_startIndexPoints.size())
         );
         glBindVertexArray(0);
 
@@ -516,6 +514,8 @@ void RenderableOrbitalKepler::render(const RenderData& data, RendererTasks&) {
     }
 
     if (renderTrails) {
+        calculateSegmentsForTrails(data);
+       
         _trailProgram->activate();
         _trailProgram->setUniform(_uniformTrailCache.opacity, opacity());
         _trailProgram->setUniform(_uniformTrailCache.color, _appearance.color);
@@ -548,12 +548,14 @@ void RenderableOrbitalKepler::render(const RenderData& data, RendererTasks&) {
         glLineWidth(_appearance.trailWidth);
 
         glBindVertexArray(_vertexArray);
+
         glMultiDrawArrays(
             GL_LINE_STRIP,
-            _si,
-            _ss,
-            static_cast<GLsizei>(_startIndex.size())
+            _startIndexTrails.data(),
+            _segmentSizeTrails.data(),
+            _lineDrawCount
         );
+
         glBindVertexArray(0);
 
         _trailProgram->deactivate();
@@ -561,7 +563,7 @@ void RenderableOrbitalKepler::render(const RenderData& data, RendererTasks&) {
 }
 
 void RenderableOrbitalKepler::updateBuffers() {
-    std::vector<kepler::Parameters> parameters = kepler::readFile(_path.value(), _format);
+    parameters = kepler::readFile(_path.value(), _format);
 
     _numObjects = parameters.size();
 
@@ -613,25 +615,27 @@ void RenderableOrbitalKepler::updateBuffers() {
         );
     }
 
-    _segmentSize.clear();
-    _startIndex.clear();
-    _startIndex.push_back(0);
-    for (size_t i = 0; i < parameters.size(); i++) {
-        const double scale = static_cast<double>(_segmentQuality) * 10.0;
-        const kepler::Parameters& p = parameters[i];
-        _segmentSize.push_back(
-            static_cast<int>(scale + (scale / std::pow(1.0 - p.eccentricity, 1.2)))
-        );
-        _startIndex.push_back(_startIndex[i] + static_cast<GLint>(_segmentSize[i]));
-    }
-    _startIndex.pop_back();
-
+    _startIndexPoints.clear();
+    _segmentSizePoints.clear();
+    _startIndexTrails.clear();
+    _segmentSizeTrails.clear();
+    _segmentSizeRaw.clear();
     size_t nVerticesTotal = 0;
-
     const int numOrbits = static_cast<int>(parameters.size());
     for (int i = 0; i < numOrbits; i++) {
-        nVerticesTotal += _segmentSize[i];
+        // For points rendering as they are always two vertices long
+        _segmentSizePoints.push_back(2);
+
+        const double scale = static_cast<double>(_segmentQuality) * 10.0;
+        const kepler::Parameters& p = parameters[i];
+        _segmentSizeRaw.push_back(
+            static_cast<int>(scale + (scale / std::pow(1.0 - p.eccentricity, 1.2)))
+        );
+        nVerticesTotal += _segmentSizeRaw[i];
     }
+    _startIndexPoints.resize(numOrbits);
+    _startIndexTrails.resize(numOrbits*2);
+    _segmentSizeTrails.resize(numOrbits*2);
     _vertexBufferData.resize(nVerticesTotal);
 
     size_t vertexBufIdx = 0;
@@ -650,9 +654,10 @@ void RenderableOrbitalKepler::updateBuffers() {
         d.setValue("Epoch", orbit.epoch);
         KeplerTranslation keplerTranslator = KeplerTranslation(d);
 
-        for (GLint j = 0 ; j < (_segmentSize[orbitIdx]); j++) {
+        const int nSegments = _segmentSizeRaw[orbitIdx];
+        for (GLint j = 0 ; j < nSegments; j++) {
             const double timeOffset = orbit.period *
-                static_cast<double>(j) / static_cast<double>(_segmentSize[orbitIdx] - 1);
+                static_cast<double>(j) / static_cast<double>(nSegments - 1);
 
             const glm::dvec3 position = keplerTranslator.position({
                 {},
@@ -660,15 +665,14 @@ void RenderableOrbitalKepler::updateBuffers() {
                 Time(0.0)
             });
 
-            _vertexBufferData[vertexBufIdx].x = static_cast<float>(position.x);
-            _vertexBufferData[vertexBufIdx].y = static_cast<float>(position.y);
-            _vertexBufferData[vertexBufIdx].z = static_cast<float>(position.z);
-            _vertexBufferData[vertexBufIdx].time = static_cast<float>(timeOffset);
-            _vertexBufferData[vertexBufIdx].epoch = orbit.epoch;
-            _vertexBufferData[vertexBufIdx].period = orbit.period;
-
-            vertexBufIdx++;
+            _vertexBufferData[vertexBufIdx + j].x = static_cast<float>(position.x);
+            _vertexBufferData[vertexBufIdx + j].y = static_cast<float>(position.y);
+            _vertexBufferData[vertexBufIdx + j].z = static_cast<float>(position.z);
+            _vertexBufferData[vertexBufIdx + j].time = static_cast<float>(timeOffset);
+            _vertexBufferData[vertexBufIdx + j].epoch = orbit.epoch;
+            _vertexBufferData[vertexBufIdx + j].period = orbit.period;
         }
+        vertexBufIdx += nSegments;
     }
 
     glBindVertexArray(_vertexArray);
@@ -703,6 +707,97 @@ void RenderableOrbitalKepler::updateBuffers() {
         }
     }
     setBoundingSphere(maxSemiMajorAxis * 1000);
+}
+
+void RenderableOrbitalKepler::calculateSegmentsForPoints(const RenderData& data) {
+    int startVertexIndex = 0;
+    for (int i = 0; i < _segmentSizeRaw.size(); i++) {
+        // Check how far along the trail we are
+        const kepler::Parameters& orbit = parameters[i];
+        const double nRevs = (data.time.j2000Seconds() - orbit.epoch) / orbit.period;
+        double frac = nRevs - std::trunc(nRevs);
+        frac += (frac < 0.0) ? 1.0: 0.0;
+
+        // Get the closest vertex before that point
+        const int nSegments = _segmentSizeRaw[i] - 1;
+        const int offset = static_cast<int>(std::floor(frac * nSegments));
+
+        // Set start vertex ID in buffer
+        _startIndexPoints[i] = startVertexIndex + offset;
+        startVertexIndex += _segmentSizeRaw[i];
+    }
+}
+
+void RenderableOrbitalKepler::calculateSegmentsForTrails(const RenderData& data) {
+    const float fade = pow(_appearance.trailFade.maxValue() - _appearance.trailFade, 2.f);
+    const double threshold = 1.0 - pow(0.05f, 1.f / fade);
+
+    int nTotalTrailParts = 0;
+    int startVertexIndex = 0;
+    for (int i = 0; i < _segmentSizeRaw.size(); i++) {
+        // Check how far along the trail we are
+        const kepler::Parameters& orbit = parameters[i];
+        const double nRevs = (data.time.j2000Seconds() - orbit.epoch) / orbit.period;
+        double frac = nRevs - std::trunc(nRevs);
+        frac += (frac < 0.0) ? 1.0 : 0.0;
+
+        int p0Start = 0;
+        int p0Length = 0;
+        int p1Start = 0;
+        int p1Length = 0;
+
+        const int nVerts = _segmentSizeRaw[i];
+        const int nSegments = nVerts - 1;
+        const int trailLength = static_cast<int>(std::ceil(threshold * nSegments));
+        if (trailLength == nSegments) {
+            // Whole trail should be visible
+            p0Start = startVertexIndex;
+            p0Length = nVerts;
+        }
+        else {
+            const int headOffset = static_cast<int>(std::ceil(frac * nSegments));
+            const int headVertexIndex = startVertexIndex + headOffset;
+            const int correctTrailLength = trailLength + 2;
+
+            int correctVertexIndex = headVertexIndex - correctTrailLength + 1;
+
+            // If the start of the trail should be at the end of the orbit
+            if (correctVertexIndex < startVertexIndex) {
+                correctVertexIndex += nVerts;
+            }
+
+            // If the trail is length passes over the last point of the orbit
+            const int lastVertexIndex = startVertexIndex + nVerts;
+            if (correctVertexIndex + correctTrailLength > lastVertexIndex) {
+                p1Start = correctVertexIndex - 1;
+                p1Length = lastVertexIndex - correctVertexIndex + 1;
+                p0Start = startVertexIndex;
+                p0Length = correctTrailLength - p1Length + 1;
+            }
+            else {
+                // If the entire trail is within the bounds of the orbit
+                p0Start = correctVertexIndex;
+                p0Length = correctTrailLength;
+            }
+        }
+
+        int newTrailParts = 0;
+        if (p0Length > 1) {
+            _startIndexTrails[nTotalTrailParts] = p0Start;
+            _segmentSizeTrails[nTotalTrailParts] = p0Length;
+            newTrailParts += 1;
+        }
+
+        if (p1Length > 1) {
+            _startIndexTrails[nTotalTrailParts + newTrailParts] = p1Start;
+            _segmentSizeTrails[nTotalTrailParts + newTrailParts] = p1Length;
+            newTrailParts += 1;
+        }
+
+        startVertexIndex += nVerts;
+        nTotalTrailParts += newTrailParts;
+    }
+    _lineDrawCount = static_cast<GLsizei>(nTotalTrailParts);
 }
 
 } // namespace openspace