… modules/space/shaders/debrisVizTrails_gs.glsl