## Info ##
This PR contains the following changes to `renderableOrbitalKepler`:
- Precomputes relevant vertices on CPU rather than filtering all vertices on the GPU.
- Removes geometry shader for trails rendering
- Changed GUI name of RenderingMode "Points+Trails" to "Points and Trails"

The changes are not breaking changes and should not affect the visuals either.

## Performance ##

The performance improvements varies based on the datasets as performance is limited based on things such as number of orbits, number of vertices per orbit, how long the trail is and how much overlap/overdraw there is for the trails/points.

Below are some before and after numbers based on asteroid assets.

### Trails ###
- Inner main, Main, Outer main:
  - Before: 80 fps
  - After: 90 fps
- Transneptunian:
  - Before: 29 fps
  - After 300+ fps
- Amor, Apollo, Outer main, Main, Inner main, Jupiter trojan, Transneptunian:
  - Before: 14 fps
  - After: 65 fps

### Points ###
- Inner main, Main, Outer main:
  - Before: 80 fps
  - After: 88 fps
- Transneptunian:
  - Before: 30 fps
  - After 300+ fps
- Amor, Apollo, Outer main, Main, Inner main, Jupiter trojan, Transneptunian:
  - Before: 14 fps
  - After: 65 fps
 
### Trails and Points ###
- Inner main, Main, Outer main:
  - Before: 42 fps
  - After: 50 fps
- Transneptunian:
  - Before: 15 fps
  - After 300+ fps
- Amor, Apollo, Outer main, Main, Inner main, Jupiter trojan, Transneptunian:
  - Before: 7 fps
  - After: 34 fps